So, seems like the videoview has a life of it's own and you really can't stop it from starting the video when it feels it has buffered enough.
So what I do now is just display the buffering dialog when the video is not playing. As the videoview/controller starts it on its own when it feels ready this should be sufficient unless we want to control the buffer.

--HG--
extra : convert_revision : svn%3Af2d6d1aa-d9ff-11de-a65c-977ad3277207/trunk%408